+++ src/controllers/ProofsController.ts
@@ -4,6 +4,7 @@ import Block from '../models/Block';
 import Leaf from '../models/Leaf';
 import { AuthUtils } from '../services/AuthUtils';
 import { BlockStatus } from '../types/BlockStatuses';
+import StatsDClient from '../lib/StatsDClient';
 
 @injectable()
 class ProofsController {
@@ -14,19 +15,26 @@ class ProofsController {
   }
 
   index = async (request: Request, response: Response): Promise<void> => {
-    const apiKeyVerificationResult = await this.authUtils.verifyApiKeyFromAuthHeader(request.headers.authorization);
+    const apiKeyVerificationResult = await this.authUtils.verifyApiKey(request, response);
 
     if (!apiKeyVerificationResult.apiKey) {
-      response.status(401).send({ error: apiKeyVerificationResult.errorMessage });
       return;
     }
 
-    const block = await Block.findOne({ status: BlockStatus.Finalized }).sort({ height: -1 }).limit(1);
+    StatsDClient?.increment('sanctuary.proofs-controller.index', undefined, {
+      projectId: apiKeyVerificationResult.apiKey.projectId,
+    });
+
+    const last3 = await Block.find({ status: BlockStatus.Finalized }).sort({ blockId: -1 }).limit(3);
+    const block = await Block.findOne({ status: BlockStatus.Finalized }).sort({ blockId: -1 }).limit(1);
+
+    console.log('last3', last3);
+    console.log(block);
 
     if (block) {
       const keys = (request.query.keys || []) as string[];
-      const leaves = await Leaf.find({ blockId: block.id, key: { $in: keys } });
-      response.send({ data: { block: block, keys: keys, leaves: leaves } });
+      const leaves = await Leaf.find({ blockId: block.blockId.toString(), key: { $in: keys } });
+      response.send({ data: { block, keys, leaves } });
     } else {
       response.send({ data: {} });
     }