+++ src/types/BlockFromPegasus.ts
@@ -1,12 +1,17 @@
-import { HexStringWith0x } from './HexStringWith0x';
+import { HexStringWith0x } from './custom';
 
 export interface BlockFromPegasus {
   _id: string;
   timestamp: Date;
   mintedAt: Date;
-  height: number;
+  blockId: number;
   root: string;
   data: Record<string, HexStringWith0x>;
-  numericFcdKeys: Array<string>;
-  numericFcdValues: Array<number>;
+  fcdKeys: Array<string>;
+  fcdValues: Array<number>;
+  anchor: string;
+  votes: Record<string, string>;
+  power: string;
+  staked: string;
+  minter: string;
 }